Refactor: #76 GroupByCount 타입을 DTO 파일로 이동
